Split NIRSpec IFU regtests (#4639)
* Split NIRSpec IFU regtests

* Remove unused spec3 test
